completed refactor of sweep line algorithm
