+++ linesegment/sweepline_test.go
@@ -1,11 +1,8 @@
 package linesegment
 
 import (
-	"github.com/google/btree"
 	"github.com/mikenye/geom2d/options"
-	"github.com/mikenye/geom2d/point"
 	"github.com/mikenye/geom2d/types"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -99,7 +96,7 @@ func FuzzFindIntersections_Int_2Segments(f *testing.F) {
 		ok, reason := compareIntersectionResults(naiveResults, sweepResults, 1e-8)
 
 		if !ok {
-			t.Fatalf("Mismatch found!\nSegments: %v\nNaive: %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
+			t.Fatalf("Mismatch found!\nSegments: %v\nNaive:      %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
 		}
 	})
 }
@@ -134,7 +131,7 @@ func FuzzFindIntersections_Int_3Segments(f *testing.F) {
 		ok, reason := compareIntersectionResults(naiveResults, sweepResults, 1e-8)
 
 		if !ok {
-			t.Fatalf("Mismatch found!\nSegments: %v\nNaive: %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
+			t.Fatalf("Mismatch found!\nSegments: %v\nNaive:      %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
 		}
 	})
 }
@@ -173,40 +170,14 @@ func FuzzFindIntersections_Int_4Segments(f *testing.F) {
 		ok, reason := compareIntersectionResults(naiveResults, sweepResults, 1e-8)
 
 		if !ok {
-			t.Fatalf("Mismatch found!\nSegments: %v\nNaive: %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
+			t.Fatalf("Mismatch found!\nSegments: %v\nNaive:      %v\nSweep Line: %v\nReason: %v\n", segments, naiveResults, sweepResults, reason)
 		}
 	})
 }
 
-func TestDeleteSegmentsFromStatus(t *testing.T) {
-	// Define the status structure as a sorted slice
-	status := []statusItem{
-		{segment: New[float64](1, 5, 3, 1)},
-		{segment: New[float64](3, 1, 5, 5)},
-		{segment: New[float64](5, 5, 1, 1)},
-	}
-
-	// Segments to delete
-	toDelete := []LineSegment[float64]{
-		New[float64](3, 1, 5, 5),
-		New[float64](5, 5, 1, 1),
-	}
-
-	// Call deleteSegmentsFromStatus
-	newStatus := deleteSegmentsFromStatus(status, toDelete, options.WithEpsilon(1e-6))
-
-	// Expected status structure
-	expected := []statusItem{
-		{segment: New[float64](1, 5, 3, 1)},
-	}
-
-	// Assert the result
-	assert.Equal(t, expected, newStatus, "Remaining status structure should match expected")
-}
-
-// TestFindIntersections ensures that the output of the sweep line algorithm (FindIntersectionsFast) matches
+// TestFindIntersectionsFast ensures that the output of the sweep line algorithm (FindIntersectionsFast) matches
 // the output of the naïve algorithm (FindIntersectionsSlow).
-func TestFindIntersections(t *testing.T) {
+func TestFindIntersectionsFast(t *testing.T) {
 	tests := map[string]struct {
 		segments []LineSegment[int]
 	}{
@@ -402,311 +373,6 @@ func TestFindIntersections(t *testing.T) {
 
 					require.True(t, InterSectionResultsEq(actualIntersections, actualIntersectionsFromSlow, options.WithEpsilon(epsilon)))
 				})
-			}
-		})
-	}
-}
-
-func TestFindLeftmostAndRightmostSegmentAndNeighbors(t *testing.T) {
-
-	// Helper function to create a pointer to a point
-	linSegPtr := func(x1, y1, x2, y2 float64) *LineSegment[float64] {
-		l := New[float64](x1, y1, x2, y2)
-		return &l
-	}
-
-	tests := map[string]struct {
-		point                point.Point[float64]
-		UofP                 []LineSegment[float64]
-		CofP                 []LineSegment[float64]
-		statusItems          []statusItem
-		expectedSL           *statusItem
-		expectedSR           *statusItem
-		expectedSPrime       *LineSegment[float64]
-		expectedSDoublePrime *LineSegment[float64]
-	}{
-		"no neighbors with single segment": {
-			point: point.New[float64](3, 3),
-			UofP: []LineSegment[float64]{
-				New[float64](2, 2, 3, 3),
-			},
-			CofP:                 nil,
-			statusItems:          []statusItem{{segment: New[float64](2, 2, 3, 3)}},
-			expectedSL:           nil,
-			expectedSR:           nil,
-			expectedSPrime:       linSegPtr(2, 2, 3, 3),
-			expectedSDoublePrime: linSegPtr(2, 2, 3, 3),
-		},
-		"right neighbor only": {
-			point: point.New[float64](3, 3),
-			UofP:  nil,
-			CofP: []LineSegment[float64]{
-				New[float64](4, 4, 3, 3),
-			},
-			statusItems: []statusItem{
-				{segment: New[float64](4, 4, 3, 3)}, // C(p)
-				{segment: New[float64](5, 5, 3, 3)}, // Right neighbor
-			},
-			expectedSL:           nil,
-			expectedSR:           &statusItem{segment: New[float64](5, 5, 3, 3)},
-			expectedSPrime:       linSegPtr(4, 4, 3, 3),
-			expectedSDoublePrime: linSegPtr(4, 4, 3, 3),
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			sPrime, sDoublePrime, sL, sR := findLeftmostAndRightmostSegmentAndNeighbors(tc.point, tc.UofP, tc.CofP, tc.statusItems)
-
-			if tc.expectedSPrime == nil {
-				assert.Nil(t, sPrime, "Expected sPrime to be nil")
-			} else {
-				assert.NotNil(t, sPrime, "Expected sPrime to be non-nil")
-				assert.Equal(t, *tc.expectedSPrime, *sPrime, "sPrime neighbor mismatch")
-			}
-
-			if tc.expectedSDoublePrime == nil {
-				assert.Nil(t, sDoublePrime, "Expected sPrime to be nil")
-			} else {
-				assert.NotNil(t, sDoublePrime, "Expected sPrime to be non-nil")
-				assert.Equal(t, *tc.expectedSDoublePrime, *sDoublePrime, "sPrime neighbor mismatch")
-			}
-
-			// Check sL neighbor
-			if tc.expectedSL == nil {
-				assert.Nil(t, sL, "Expected sL neighbor to be nil")
-			} else {
-				assert.NotNil(t, sL, "Expected sL neighbor to be non-nil")
-				assert.Equal(t, *tc.expectedSL, *sL, "Left neighbor mismatch")
-			}
-
-			// Check sR neighbor
-			if tc.expectedSR == nil {
-				assert.Nil(t, sR, "Expected sR neighbor to be nil")
-			} else {
-				assert.NotNil(t, sR, "Expected sR neighbor to be non-nil")
-				assert.Equal(t, *tc.expectedSR, *sR, "Right neighbor mismatch")
-			}
-		})
-	}
-}
-
-func TestFindNeighbors(t *testing.T) {
-	tests := map[string]struct {
-		statusItems   []statusItem // The sorted status structure (slice of statusItem)
-		point         qItem        // The event point to find neighbors for
-		expectedLeft  *statusItem  // Expected left neighbor (nil if none exists)
-		expectedRight *statusItem  // Expected right neighbor (nil if none exists)
-	}{
-		"point with both neighbors": {
-			statusItems: []statusItem{
-				{segment: New[float64](2, 4, 1, 1)}, // Left neighbor
-				{segment: New[float64](4, 6, 3, 3)}, // Matching segment
-				{segment: New[float64](4, 4, 5, 2)}, // Right neighbor
-			},
-			point:         qItem{point: point.New[float64](3, 3)},
-			expectedLeft:  &statusItem{segment: New[float64](2, 4, 1, 1)},
-			expectedRight: &statusItem{segment: New[float64](4, 4, 5, 2)},
-		},
-		"point with no left neighbor": {
-			statusItems: []statusItem{
-				{segment: New[float64](4, 6, 3, 3)}, // Matching segment
-				{segment: New[float64](4, 4, 5, 2)}, // Right neighbor
-			},
-			point:         qItem{point: point.New[float64](3, 3)},
-			expectedLeft:  nil,
-			expectedRight: &statusItem{segment: New[float64](4, 4, 5, 2)},
-		},
-		"point with no right neighbor": {
-			statusItems: []statusItem{
-				{segment: New[float64](2, 4, 1, 1)}, // Left neighbor
-				{segment: New[float64](4, 6, 3, 3)}, // Matching segment
-			},
-			point:         qItem{point: point.New[float64](3, 3)},
-			expectedLeft:  &statusItem{segment: New[float64](2, 4, 1, 1)},
-			expectedRight: nil,
-		},
-		"point with no neighbors (single item)": {
-			statusItems: []statusItem{
-				{segment: New[float64](1, 1, 3, 3)},
-			},
-			point:         qItem{point: point.New[float64](3, 3)},
-			expectedLeft:  nil,
-			expectedRight: nil,
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-
-			t.Logf("Status Items: %v", tc.statusItems)
-			left, right := findNeighbors(tc.statusItems, tc.point)
-
-			// Check left neighbor
-			if tc.expectedLeft == nil {
-				assert.Nil(t, left, "Expected left neighbor to be nil")
-			} else {
-				assert.NotNil(t, left, "Expected left neighbor to be non-nil")
-				assert.Equal(t, *tc.expectedLeft, *left, "Left neighbor mismatch")
-			}
-
-			// Check right neighbor
-			if tc.expectedRight == nil {
-				assert.Nil(t, right, "Expected right neighbor to be nil")
-			} else {
-				assert.NotNil(t, right, "Expected right neighbor to be non-nil")
-				assert.Equal(t, *tc.expectedRight, *right, "Right neighbor mismatch")
-			}
-		})
-	}
-}
-
-func TestFindNewEvent(t *testing.T) {
-	// Define a small epsilon for geometric calculations
-	epsilon := 1e-9
-	opts := []options.GeometryOptionsFunc{options.WithEpsilon(epsilon)}
-
-	// Helper function to create a pointer to a point
-	pointPtr := func(x, y float64) *point.Point[float64] {
-		p := point.New[float64](x, y)
-		return &p
-	}
-
-	tests := map[string]struct {
-		segment1         LineSegment[float64]
-		segment2         LineSegment[float64]
-		currentPoint     point.Point[float64]
-		expectedPoint    *point.Point[float64] // Nil if no intersection is expected
-		expectedInQueue  bool                  // Whether the point should be in the queue
-		expectedCountInQ int                   // The number of matching points in the queue
-	}{
-		"valid intersection below current point": {
-			segment1:         New[float64](0, 5, 5, 0),
-			segment2:         New[float64](0, 0, 5, 5),
-			currentPoint:     point.New[float64](3, 4),
-			expectedPoint:    pointPtr(2.5, 2.5),
-			expectedInQueue:  true,
-			expectedCountInQ: 1,
-		},
-		"intersection above current point": {
-			segment1:         New[float64](0, 5, 5, 0),
-			segment2:         New[float64](0, 0, 5, 5),
-			currentPoint:     point.New[float64](2, 2),
-			expectedPoint:    nil,
-			expectedInQueue:  false,
-			expectedCountInQ: 0,
-		},
-		"no intersection": {
-			segment1:         New[float64](0, 5, 5, 5),
-			segment2:         New[float64](0, 0, 5, 0),
-			currentPoint:     point.New[float64](3, 4),
-			expectedPoint:    nil,
-			expectedInQueue:  false,
-			expectedCountInQ: 0,
-		},
-		"intersection already in queue": {
-			segment1:         New[float64](0, 5, 5, 0),
-			segment2:         New[float64](0, 0, 5, 5),
-			currentPoint:     point.New[float64](3, 4),
-			expectedPoint:    pointPtr(2.5, 2.5),
-			expectedInQueue:  true,
-			expectedCountInQ: 1, // Ensure no duplicates are added
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			// Initialize the event queue
-			Q := btree.NewG[qItem](2, qItemLess)
-
-			// Initialise results
-			R := newIntersectionResults[float64]()
-
-			// Prepopulate the queue for cases where intersection is already in queue
-			if tc.expectedPoint != nil && tc.expectedCountInQ > 0 {
-				Q.ReplaceOrInsert(qItem{
-					point:    *tc.expectedPoint,
-					segments: []LineSegment[float64]{tc.segment1, tc.segment2},
-				})
-			}
-
-			// Call findNewEvent
-			findNewEvent(tc.segment1, tc.segment2, tc.currentPoint, Q, R, opts...)
-
-			// Verify the count of matching points in the queue
-			matchingCount := 0
-			Q.Ascend(func(item qItem) bool {
-				if tc.expectedPoint != nil && item.point.Eq(*tc.expectedPoint, opts...) {
-					matchingCount++
-				}
-				return true
-			})
-			assert.Equal(t, tc.expectedCountInQ, matchingCount,
-				"Mismatch in the number of matching intersection points in the queue")
-
-			// Check if the expected point is in the queue
-			if tc.expectedPoint != nil {
-				found := Q.Has(qItem{point: *tc.expectedPoint})
-				assert.Equal(t, tc.expectedInQueue, found,
-					"Intersection point presence in queue mismatch")
-			} else {
-				// If no expected point, ensure no new event was added
-				assert.Equal(t, 0, Q.Len(), "No new intersection point should be added")
-			}
-		})
-	}
-}
-
-func TestSortStatusBySweepLine(t *testing.T) {
-	tests := map[string]struct {
-		input    []statusItem
-		sweepY   float64
-		expected []statusItem
-	}{
-		"basic case": {
-			input: []statusItem{
-				{segment: New[float64](3, 1, 5, 5)}, // Will have XAtY = 4 at sweepY = 3
-				{segment: New[float64](1, 1, 3, 3)}, // Will have XAtY = 3 at sweepY = 3
-				{segment: New[float64](2, 4, 1, 1)}, // Will have XAtY = 1.5 at sweepY = 3
-			},
-			sweepY: 3,
-			expected: []statusItem{
-				{segment: New[float64](2, 4, 1, 1)}, // XAtY = 1.5
-				{segment: New[float64](1, 1, 3, 3)}, // XAtY = 3
-				{segment: New[float64](3, 1, 5, 5)}, // XAtY = 4
-			},
-		},
-		"already sorted": {
-			input: []statusItem{
-				{segment: New[float64](2, 4, 1, 1)},
-				{segment: New[float64](1, 1, 3, 3)},
-				{segment: New[float64](3, 1, 5, 5)},
-			},
-			sweepY: 3,
-			expected: []statusItem{
-				{segment: New[float64](2, 4, 1, 1)},
-				{segment: New[float64](1, 1, 3, 3)},
-				{segment: New[float64](3, 1, 5, 5)},
-			},
-		},
-		"empty input": {
-			input:    []statusItem{},
-			sweepY:   3,
-			expected: []statusItem{},
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			sortStatusBySweepLine(tc.input, qItem{point: point.New(0, tc.sweepY)})
-
-			//debugPrintStatus(tc.input, tc.sweepY)
-
-			//fmt.Println("input: ", tc.input)
-			//fmt.Println("expect:", tc.expected)
-
-			for i := range tc.input {
-				assert.Equal(t, tc.expected[i].segment, tc.input[i].segment, "Segment mismatch at index %d", i)
 			}
 		})
 	}
